@@ -12,6 +12,6 @@ ActiveAdmin.register TestResult do
   #   permitted << :other if params[:action] == 'create' && current_user.admin?
   #   permitted
   # end
-  permit_params :well_id, :test_id, :value
+  permit_params :well_id, :test_id, :state
 
 end
