@@ -12,6 +12,6 @@ ActiveAdmin.register Test do
   #   permitted << :other if params[:action] == 'create' && current_user.admin?
   #   permitted
   # end
-  permit_params :plate_id, :user_id, :value
+  permit_params :plate_id, :user_id
 
 end
