@@ -12,6 +12,6 @@ ActiveAdmin.register User do
   #   permitted << :other if params[:action] == 'create' && current_user.admin?
   #   permitted
   # end
-  permit_params :dob, :telno, :email, :role, :confirmed_at
+  permit_params :email, :role, :confirmed_at, :password, :password_confirmation, :security_question_id, :reset_password_token, :reset_password_sent_at, :remember_created_at, :confirmation_token, :confirmation_sent_at, :api_key, :locked_at, :failed_attempts, :unlock_token, :security_question_answer, :is_active
   
 end
