@@ -12,8 +12,13 @@
 //= require bs-custom-file-input/dist/bs-custom-file-input.min.js
 //= require chart.js/dist/Chart.min.js
 //= require jsqr/dist/jsQR.js
-//= require jquery
-//= require jquery-ui/widgets/autocomplete
+//= require jquery/dist/jquery.min.js
+//= require jquery-ui-dist/jquery-ui.min.js
+//= require bootstrap-select/dist/js/bootstrap-select.min.js
+//= require datatables.net/js/jquery.dataTables.min.js
+//= require datatables.net-bs4/js/dataTables.bootstrap4.min.js
+//= require daterangepicker/moment.min.js
+//= require daterangepicker/daterangepicker.js
 //= require popper
 //= require bootstrap-sprockets
 //= require rails-ujs
