@@ -16,6 +16,7 @@ module OpencellTesting
     # Application configuration can go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded after loading
     # the framework and any gems in your application.
-    Time.zone = 'London'
+    config.time_zone = 'London'
+    config.generators.javascript_engine :js
   end
 end
