@@ -1,6 +1,6 @@
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
-
+  config.web_console.allowed_ips = '172.0.0.0/8'
   # In the development environment your application's code is reloaded on
   # every request. This slows down response time but is perfect for development
   # since you don't have to restart the web server when you make code changes.
@@ -12,7 +12,6 @@ Rails.application.configure do
   # Show full error reports.
   config.consider_all_requests_local = true
 
-  config.email_test_results = false
 
   # Enable/disable caching. By default caching is disabled.
   # Run rails dev:cache to toggle caching.
@@ -39,6 +38,9 @@ Rails.application.configure do
   config.action_mailer.perform_caching = false
   config.action_mailer.default_url_options = { :host => 'localhost:3000' }
 
+  config.active_job.queue_adapter     = :sidekiq
+  # config.active_job.queue_name_prefix = "opencell_testing_development"
+  
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
 
