@@ -1,10 +1,12 @@
 FactoryBot.define do
   factory :plate do
     state { Plate.states[:preparing] }
-    # transient do
-    #   well_count  {96}
-    # end
-    #
+    lab { association :lab }
+    user { create(:user) }
+    transient do
+      well_count  {96}
+    end
+    
     # after(:build) do |plate, evaluator|
     #   create_list(:well, evaluator.well_count, plate: plate)
     # end
